ttrss-mobile using a HTML5 application cache [1] for all of its assets.
This has the benifit of the client keeping a preloaded local cache of
assets thus speeding up launch of the application as it only requires
a single call to ttrss-mobile.appcache to see if it has been modified
(usally not, so we get a 304).  I've included a python script
updateAppCache.py to build the manifest file up for you when a change
an asset or do a version release.  Additonally there is a .htaccess
file to set the proper content-type for the manifest file.

[1] http://www.html5rocks.com/en/tutorials/appcache/beginner/